Use ec-automation account and sign the commit
Worth a try, I can see that the commits that do not bring up the
`Red Hat Konflux` check have the GitHub App identity whereas we have
the `$GITHUB_ACTOR`, which might not be the GitHub App identity at all
times (e.g. when manually run). Also, the commits are signed using
GitHub's GPG key. We can't access that one, so perhaps using the SSH key
might do the trick.